+++ onsiteprint-plugin/CHANGELOG.md
@@ -22,6 +22,12 @@ This document includes the change log notes for the OnsitePrint Plugin - [GitHub
 
     ---
 
+- Added New `Error Title` and `Error Description` to Modal, used in /modal.php.
+    
+    [EB | block-template.js](blocks/event/block-template.php)
+
+    ---
+
 - Added new Script `Parts`.
 - Added new Function, `opAddParticipant()`.
 - Added new Function, `opSetupHeader()`.